+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.1
+
+* Add the `Crypto.Cipher.ChaChaPoly1305.Conduit` and
+  `Crypto.PubKey.ECIES.Conduit` modules
+
 ## 0.2.0
 
 * Add HMAC sink